Credentials_test.py
import unittest
from Credentials import Credentials
class TestUnit(unittest.TestCase):
    '''
    Test class that defines test cases for the Credentials class behaviours.
    Args:
    unittest.TestCase:TestCase class that helps in creating test cases
    '''
    def setUp(self):
        '''
        Set up method to run before each test cases.
        '''
        self.new_Credentials=Credentials("instagram","angelakarenzi5","password")

    def test_init(self):
        '''
        test_init test case to test if the object is initialized properly
        '''

        self.assertEqual(self.new_Credentials.website_name,"instagram")
        self.assertEqual(self.new_Credentials.user_name,"angelakarenzi5")
        self.assertEqual(self.new_Credentials.password,"password")
    

    def test_save_Credentials(self):
        '''
        test_save_Credentials test case to test if the Credentials object is saved into
        the Credentials list
        '''
        self.new_Credentials.save_Credentials() # saving the new Credentials
        self.assertEqual(len(Credentials.Credentials_list),1)

    # Items up here...

    def test_save_multiple_Credentials(self):
        '''
        test_save_multiple_Credentials to check if we can save multiple Credentials
        objects to our Credentials_list
        '''
        self.new_Credentials.save_Credentials()
        test_Credentials = Credentials("instagram","angelakarenzi5","password") # new Credentials
        test_Credentials.save_Credentials()
        self.assertEqual(len(Credentials.Credentials_list),3)

    # setup and class creation up here

    def tearDown(self):
            '''
            tearDown method that does clean up after each test case has run.
            '''
            Credentials.Credentials_list = []

    # other test cases here
    def test_save_multiple_Credentials(self):
            '''
            test_save_multiple_Credentials to check if we can save multiple Credentials
            objects to our Credentials_list
            '''
            self.new_Credentials.save_Credentials()
            test_Credentials = Credentials("instagram","angelakarenzi5","password") # new credentials
            test_Credentials.save_Credentials()
            self.assertEqual(len(Credentials.Credentials_list),2)

    # More tests above
    def test_delete_Credentials(self):
            '''
            test_delete_Credentials to test if we can remove a credentials from our credentials list
            '''
            self.new_Credentials.save_Credentials()
            test_Credentials = Credentials("instagram","angelakarenzi5","password") # new Credentials
            test_Credentials.save_Credentials()

            self.new_Credentials.delete_Credentials()# Deleting a Credentials object
            self.assertEqual(len(Credentials.Credentials_list),1)

    def test_find_credentials_by_name(self):
        '''
        test to check if we can find a Credentials by phone name and display information
        '''

        self.new_Credentials.save_Credentials()
        self.new_Credentials.save_Credentials()
        test_Credentials = Credentials("instagram","angelakarenzi5","password") # new Credentials
        test_Credentials.save_Credentials()

        found_Credentials = Credentials.find_by_name("instagram")

        self.assertEqual(found_Credentials.user_name,test_Credentials.user_name)

    def test_display_all_Credentials(self):
        '''
        method that returns a list of all Credentials saved
        '''

        self.assertEqual(Credentials.display_Credentials(),Credentials.Credentials_list)


    
if __name__ ==  '__main__':
    unittest.main()
